module2/data_processing_pipeline/requirements.txt
black==25.12.0
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
docker==7.1.0
idna==3.11
iniconfig==2.3.0
librt==0.7.4
loguru==0.7.3
mypy==1.19.1
mypy_extensions==1.1.0
packaging==25.0
pathspec==0.12.1
platformdirs==4.5.1
pluggy==1.6.0
psycopg2-binary==2.9.11
Pygments==2.19.2
pytest==9.0.2
pytest-mock==3.15.1
python-dotenv==1.2.1
pytokens==0.3.0
pywin32==311
requests==2.32.5
ruff==0.14.9
testcontainers==4.13.3
typing_extensions==4.15.0
urllib3==2.6.2
win32_setctime==1.2.0
wrapt==2.0.1
